refactor: 필터링 로직 hook으로 이동
