+++ app/(main)/schedule/hooks/useSchedules.ts
@@ -1,11 +1,19 @@
 import { useMemo } from "react";
-import { Day, Mentor } from "@/types/mentor";
+import { useGetReservedSchedules } from "@/apis/mentor/hooks/useGetReservedSchedules";
+import { Day } from "@/types/mentor";
+import { toYYYYMMDD } from "@/utils/dateUtils";
 import { toHHMM } from "@/utils/time";
 import { createTimeRangeList, getDisabledDays } from "../utils/scheduleUtils";
 
-export const useSchedules = (schedules: NonNullable<Mentor["schedules"]>) => {
+export const useSchedules = (mentorId: number, currentDate: Date) => {
+  const { data } = useGetReservedSchedules(mentorId, {
+    year: currentDate?.getFullYear(),
+    month: currentDate?.getMonth() + 1,
+  });
+  const { schedules, reserved } = data ?? {};
+
   const disabledDays = useMemo(
-    () => getDisabledDays(schedules.map(({ dayOfWeek }) => dayOfWeek)),
+    () => getDisabledDays(schedules?.map(({ dayOfWeek }) => dayOfWeek) ?? []),
     [schedules]
   );
 
@@ -19,5 +27,19 @@ export const useSchedules = (schedules: NonNullable<Mentor["schedules"]>) => {
     );
   }, [schedules]);
 
-  return { disabledDays, timeRangeListPerDay };
+  const currentDay = new Intl.DateTimeFormat("ko-KR", {
+    weekday: "short",
+  }).format(currentDate) as Day;
+
+  const availableTimeRangeList = useMemo(
+    () =>
+      timeRangeListPerDay?.[currentDay]?.filter((timeRange) => {
+        return !reserved?.[toYYYYMMDD(currentDate)]?.some(
+          ([start, end]) => start.startsWith(timeRange[0]) && end.startsWith(timeRange[1])
+        );
+      }),
+    [currentDate, currentDay, reserved, timeRangeListPerDay]
+  );
+
+  return { disabledDays, availableTimeRangeList };
 };
